For beta testers only, makes beta compatible with new ERCF V2
Merry Christmas to all Beta Team Testers.

This is the only part that needs to be update to make your current version work with the recently released ERCF V2 work (I haven't tested it so print one to be sure it fits). It should line up just fine, and the switches are in the correct position for it to work with the latest version of Happy Hare.

The older version with switches are not compatible with the latest version of Happy Hare, so if you do have that version, either disable the switches or print this new version.

Thanks again to all beta testers and happy holidays!